routes: split God module into per-resource files (#187 first slice) (v5.95.1)
`aifw-api/src/routes.rs` had grown to 4013 lines with 105 public
handlers covering ~13 distinct API resources. Per the issue's plan,
this commit converts it from a single file to a `routes/` module
directory and extracts the cleanest, most-bounded resources as a
proof-of-pattern. Remaining resources can move incrementally in
follow-up commits without disturbing the layout established here.

What landed:

- `routes.rs` → `routes/mod.rs` (via `git mv` so blame is preserved
  through the rename).
- Shared helpers (`bad_request`, `internal`, `port_range`) bumped to
  `pub(super)` so sub-modules can use them.
- `routes/geoip.rs` — `CreateGeoIpRuleRequest` + 5 handlers
  (list/create/update/delete + lookup), ~95 lines.
- `routes/schedules.rs` — `Schedule`, `StringOrVec`,
  `CreateScheduleRequest`, validation helpers, + 4 handlers, ~190
  lines.
- `routes/dns.rs` — `DnsConfigRequest`, `DnsConfigResponse`, +
  `get_dns`/`update_dns`, ~70 lines.
- mod.rs re-exports each sub-module's items (`pub use geoip::*` etc.)
  so existing `routes::handler_name` call sites in main.rs and tests
  keep working without per-handler import churn during the migration.

Result: mod.rs is now 3689 lines (down 324 from 4013) with three
focused sibling modules taking the extracted concerns. The same
pattern fits the remaining resources (rules, nat, vpn, static_routes,
status, settings, auth, users, backup, interfaces); doing them all
in one PR would be high-churn and high-risk, so the issue stays open
for the rest.

Verified:
- `cargo check --workspace` clean
- `cargo test --package aifw-api` — 123 tests pass
