+++ aifw-ui/package.json
@@ -1,6 +1,6 @@
 {
   "name": "aifw-ui",
-  "version": "5.95.0",
+  "version": "5.95.1",
   "private": true,
   "scripts": {
     "dev": "next dev",
